Refactor entity IDs from Integer to Long and enhance MovieService with pagination support

+++ src/main/java/com/ndduroc/rocmovies/Entity/Borrow.java
@@ -17,7 +17,7 @@ import jakarta.validation.constraints.Size;
 public class Borrow {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer borrowId;
+    private Long borrowId;
     
     @Temporal(TemporalType.DATE)
     private Date date;
